+++ token/core/fabtoken/v1/validator/validator_transfer.go
@@ -17,6 +17,7 @@ import (
 	"github.com/LFDT-Panurus/panurus/token/services/identity"
 	htlc2 "github.com/LFDT-Panurus/panurus/token/services/identity/interop/htlc"
 	"github.com/LFDT-Panurus/panurus/token/services/interop/htlc"
+	"github.com/LFDT-Panurus/panurus/token/services/utils"
 	"github.com/LFDT-Panurus/panurus/token/token"
 	"github.com/hyperledger-labs/fabric-smart-client/pkg/utils/errors"
 )
@@ -50,7 +51,7 @@ func TransferSignatureValidate(c context.Context, ctx *Context) error {
 			}
 			verifierCache[ownerKey] = verifier
 		}
-		if common.IsNil(ctx.SignatureProvider) {
+		if utils.IsNil(ctx.SignatureProvider) {
 			return common.ErrNilSignatureProvider
 		}
 		ctx.Logger.Debugf("signature verification [%v][%s]", tok, driver.Identity(owner).UniqueID())
