+++ token/core/zkatdlog/nogh/v1/validator/validator_issue.go
@@ -13,6 +13,7 @@ import (
 	"github.com/LFDT-Panurus/panurus/token/core/common"
 	"github.com/LFDT-Panurus/panurus/token/core/zkatdlog/nogh/v1/issue"
 	"github.com/LFDT-Panurus/panurus/token/services/logging"
+	"github.com/LFDT-Panurus/panurus/token/services/utils"
 	"github.com/hyperledger-labs/fabric-smart-client/pkg/utils/errors"
 )
 
@@ -55,7 +56,7 @@ func IssueValidate(c context.Context, ctx *Context) error {
 	if err != nil {
 		return errors.Wrapf(err, "failed getting verifier for issuer [%s]", action.Issuer.String())
 	}
-	if common.IsNil(ctx.SignatureProvider) {
+	if utils.IsNil(ctx.SignatureProvider) {
 		return common.ErrNilSignatureProvider
 	}
 	if _, err := ctx.SignatureProvider.HasBeenSignedBy(c, action.Issuer, verifier); err != nil {
