+++ token/services/utils/functions.go
@@ -7,6 +7,7 @@ SPDX-License-Identifier: Apache-2.0
 // This file provides generic utility functions including identity functions and nil checking.
 // IdentityFunc returns a function that returns its input unchanged.
 // IsNil checks if a value is nil using reflection for pointer-like types.
+
 package utils
 
 import "reflect"
@@ -15,9 +16,16 @@ func IdentityFunc[T any]() func(T) T {
 	return func(t T) T { return t }
 }
 
-func IsNil[T any](value T) bool {
-	// Use reflection to check if the value is nil
-	v := reflect.ValueOf(value)
-
-	return (v.Kind() == reflect.Pointer || v.Kind() == reflect.Slice || v.Kind() == reflect.Map || v.Kind() == reflect.Chan || v.Kind() == reflect.Func || v.Kind() == reflect.Interface) && v.IsNil()
+// IsNil returns true for nil values and interfaces containing typed nil values.
+func IsNil(value any) bool {
+	if value == nil {
+		return true
+	}
+	reflected := reflect.ValueOf(value)
+	switch reflected.Kind() {
+	case reflect.Chan, reflect.Func, reflect.Interface, reflect.Map, reflect.Pointer, reflect.Slice:
+		return reflected.IsNil()
+	default:
+		return false
+	}
 }
